+++ src/transporte/Transporte.java
@@ -3,9 +3,9 @@ package transporte;
 public abstract class Transporte {
 	
 	private String patente;
-	private double velocidad;
+	private float velocidad;
 	
-	public Transporte(String patente, double velocidad) {
+	public Transporte(String patente, float velocidad) {
 		super();
 		this.patente = patente;
 		this.velocidad = velocidad;
@@ -18,10 +18,10 @@ public abstract class Transporte {
 		this.patente = patente;
 	}
 
-	public double getVelocidad() {
+	public float getVelocidad() {
 		return velocidad;
 	}
-	public void setVelocidad(double velocidad) {
+	public void setVelocidad(float velocidad) {
 		this.velocidad = velocidad;
 	}
 	
@@ -37,13 +37,13 @@ public abstract class Transporte {
 
 	public void setOcupadoComun(int ocupadoComun) {}
 	
-	public double getPrecioKM(){ return 0;}
+	public float getPrecioKM(){ return 0;}
 	/**
 	 * getPrecioKM() devuelve el precio x km, en los casos que corresponda, lo
 	 * multiplica por la cantidad de pasakeros (todos menos auto)
 	 */
 	
-	public double getCosto(){return 0;}
+	public float getCosto(){return 0;}
 	/**
 	 * getCosto() devuelve el precio base, en caso de que no lo haya devuelve 0 (en colectivos)
 	 */
